+++ CHANGELOG.md
@@ -6,7 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
-Nothing yet.
+_Changes since 1.7.3_
+
+### Added
+- A "Save Raw" button at the response viewer to save raw response body to a file
 
 
 ## [1.7.3] -- 2025-04-29